Clean up imports and fingerprinting.
