+++ aleph/model/common.py
@@ -1,7 +1,10 @@
 import uuid
 import string
 import fingerprints
+from hashlib import sha1
 from datetime import datetime
+from sqlalchemy import func
+from sqlalchemy.dialects.postgresql import ARRAY
 
 from aleph.core import db
 from aleph.text import string_value
@@ -10,24 +13,47 @@ from aleph.text import string_value
 ALPHABET = string.ascii_lowercase + string.digits
 
 
-def make_token():
-    num = uuid.uuid4().int
-    s = []
-    while True:
-        num, r = divmod(num, len(ALPHABET))
-        s.append(ALPHABET[r])
-        if num == 0:
-            break
-    return ''.join(reversed(s))
-
-
 def make_textid():
     return uuid.uuid4().hex
 
 
-def make_fingerprint(text, **kwargs):
-    """Generate a normalised entity name, used for the graph."""
-    return fingerprints.generate(string_value(text))
+def object_key(obj):
+    """Generate a checksum for a nested object or list."""
+    key = sha1()
+
+    if isinstance(obj, (list, set, tuple)):
+        for o in obj:
+            o = object_key(o)
+            if o is not None:
+                key.update(o)
+    elif isinstance(obj, dict):
+        for k, v in obj.items():
+            v = object_key(v)
+            if v is not None:
+                key.update(k)
+                key.update(v)
+    else:
+        v = string_value(obj)
+        if v is not None:
+            key.update(v.encode('utf-8'))
+
+    return key.hexdigest()
+
+
+def merge_data(base, merge):
+    """Merge two objects such that values in base are kept
+    and updated only if merge has additional info."""
+    if isinstance(base, (list, set, tuple)):
+        data = base + merge
+        data = {object_key(d): d for d in data}
+        return data.values()
+    if isinstance(base, dict):
+        data = dict(base)
+        for k, v in merge.items():
+            b = base.get(k, v)
+            data[k] = merge_data(b, v)
+        return data
+    return merge if base is None else base
 
 
 class IdModel(object):
@@ -112,3 +138,16 @@ class SoftDeleteModel(DatedModel):
         data = parent.to_dict() if hasattr(parent, 'to_dict') else {}
         data['deleted_at'] = self.deleted_at
         return data
+
+
+class ModelFacets(object):
+
+    @classmethod
+    def facet_by(cls, q, field, filter_null=False):
+        if isinstance(field.property.columns[0].type, ARRAY):
+            field = func.unnest(field)
+        cnt = func.count(field)
+        q = q.from_self(field, cnt)
+        q = q.group_by(field)
+        q = q.order_by(cnt.desc())
+        return [{'value': v, 'count': c} for v, c in q if v is not None]
